Added 'sleep 3' workaround.
When used, it reduces the time for compile the
test-suite from 23 minutes to 2 minutes. It also
reduces plain swig compilation

This is how it is used:

1.- copy or link the Tools/sleep program
    into a directory in your path, like ${HOME}/bin.

2.- check the right path for the real sleep program,
    now it says

       SLEEP=/bin/sleep

    but in your system could be different

3.- compile like this

      NOSLEEP=1 make
      NOSLEEP=1 make check-python-test-suite

    or

      export NOSLEEP=1
      make
      make check-python-test-suite

    just rembember to unset NOSLEEP if you are
    not using it anymore.

I don't see any problem in the compilation
if sleep is deactivated. But if in your
system/platform the compilation fails,
just don't set NOSLEEP, and the compilation
will go back to sleep as usual again.


This is just a work around until the the 'sleep 3'
problem is further analized.

Marcelo