Correct explanation of how to match on class name in %rename.
Replace incorrect documentation of $parentNode from %rename discussion: it
advised using match$parentNode but this doesn't work because the parent node
is not yet set when %rename is parsed.

Document the "fullname" attribute of %rename which can be used to restrict the
match to the given full name of a declaration only.